Create BaseController and BaseModel
